+++ tests/mosaic/BUILD
@@ -54,6 +54,16 @@ jax_py_test(
     ] + py_deps("absl/testing"),
 )
 
+jax_py_test(
+    name = "gpu_layout_inference_test",
+    srcs = ["gpu_layout_inference_test.py"],
+    deps = [
+        "//jax",
+        "//jax:mosaic_gpu",
+        "//jax:test_util",
+    ] + py_deps("absl/testing"),
+)
+
 jax_multiplatform_test(
     name = "matmul_test",
     srcs = ["matmul_test.py"],
